Change enum behaviour:
Enums now get completely expanded and any items not present in a dynamic
enum are filled by separators. This makes it possible to have
predictable positions even if some items are missing.

The iteration code expects that the dynamic and initial enum lists have
the same ordering of items, which is not unreasonable.