Separate property for viewport shade pies.
This is fine tuned for using with pie menus, where solid/wire mode
is given the most important west/east positions.

There were two possibilities with supporting custom pie enums for an
existing property. One is to add an extra function that dynamically
fills the pie items to fit a pie scheme. The other is make a separate
python property. I feel this is the correct way because it does not
burden all enum properties with extra function storage and it does not
always need to be dynamic either.

Also cleanup minor debug line