OpenSubdiv: Update GLSL code to use geometry shader
This bumps GLSL requirement for the GPU, but this is the only way
we can support fare-vertex attributes. Shader currently supports
solid draw only but it've got lighting model hooked up.

This gives quite noticeable slowdown on my gt620, but this card
is not really something powerful and likely code could be improved
later.

This also allowed to have proper normals for flat shading.

For now the code stays disabled for until some runtime check or
so is implemented.

TODO: There are some matrix functions which are currently simply
copy-pasted from the blenlib, would be nice to clean this up.