OpenSubdiv: Initial work towards UV maps support
This lands basic test of pass UV coordinates to the coarse mesh
via evaluator C-API and implements all the needed interpolation
in the GLSL shader.

This code is not used still because GLSL shader is not in use
because enabling it will make it impossible to run anything on
the older cards.

Still need to do some runtime check for this..