OpenSubdiv: Reduce number of glDrawEleemnts calls when in solid view
Made it so OsdGLMesh partitions are drawn in batches, so if mesh does have
a continuous set of faces with the same material they all will be displayed
with a single glDrawEelemnts call.

Not sure how this affects on performance on the better quality cards, but
on laptop it brings FPS back to the same value as it was before implementing
multiple textures in solid view (it's around 25% speed gain on Intel card).

That's a good practice anyway to reduce number of glDrawElements anyway, so
even if we wouldn't support low-end GPU with old driver (because of geometry
shader) it's still good to have.

Improvements to textured and GLSL display will be done later,