OpenSubdiv: Initial glue to the GPU_material routines
For now this only tested on matcaps, the rest of GLSL display is
considered unworkable.

This implements ability to define geometry shader for the GPU
material which later is to be used for texture coordinates generation.

For now GOU materials are split depending on thether they're used
for OpenSubdiv mesh display or for "regular" mesh display. In the
future it'll be possible to get rid of such a split, but that'd
require a bit of changes in how this stuff is used all over the
place. better to do it after the viewport refactor project.

And we're now officially only supporting GPU-side tessellation
only for rather modern GPU, supporting older ones becomeng more
and more difficult. If GPU does not support recent GLSL shading
then Blender will fallback to CPU-side grids evaluation, but it
will still happen using OpenSubdiv. So in the future (once we'll
support uniform subdivisions in evlauator API) there'll be no
differences between old and new GPU visualization.

For those who're using older GPU added new compute type in user
preferences which is called NONE. This means CPU-side tessellation
will be used.