OpenSubdiv: Prevent garbage shading when displaying vertex colors/weights
This is for cases when vertex colors or weights are visualising. Currently
we don't have CD layers for this data and here we only make it so there's
no garbage displayed.

In the future we'll either need to have CD for this data or pass this data
as face-varying or vertex-varying data in OSD mesh.