Added a "select" property to GPencil strokes
This property doesn't serve much purpose yet, but will soon be used to tag strokes
for modification by "in-viewport" stroke editing tools.

As a temporary visualisation hack, this uses the abandoned "debug" drawing mode
code, which plots out all the points of the stroke. We might want to replace this
with something nicer later, but this will do for now.