Tracking: Fix crashes and some spontaneous track loss
Crash was caused by the fact that requested region might go outside
of the actual frame. For now we deal with this by only filling the
region partially with the data we can and keep rest of the image
empty.

There also used to be a bug in the detection that tracking failed
caused by reading uninitialized variable.