Added a dedicated way of drawing selected strokes and their points
This currently uses the theme setting for vertex size for the selected colour of
strokes. For the unselected colour, it still uses the colour of the stroke itself,
which only works well when the strokes are not fully opaque, and also if the
colour of the stroke is not the same as the selected colour.