Bring back image input, checker, viewer and output nodes
This more like a prove-of-concept change because well, we've got
a dependency solver in the nodes already which seems to work just
fine for the shading nodes. So want to see if this is something we
can use for texture nodes now.

This isn't that far away from the original idea of having some
SVM-like machine for texture nodes.

Hell of a lot of nodes are not gonna to work yet,