Cleanup: Remove transformation matrix classes
Not sure why depsgraph should actually care about transformation,
and for sure it's kinda weird to have wrappers. It's a blender core
and better to avoid using wrappers/abstractions in here.

The idea is to not store data in the depsgraph anyway, but use an
evaluation context for this instead.