GP Onion Skinning: Custom before/after colours and ranges
It is now possible to set different colors for ghosts "before" and "after" the
current frame, and to separately control how many are displayed on each side.

To use custom colors, it is necessary to firstly toggle the color-wheel icon-toggle
(located beside the Onion Skinning) checkbox. This will make both before and after
strokes get drawn using custom colors. I tried having separate toggles, but that
was too clumsy.


TODO:
* Some version-patching code will be needed to initialise the colours and fields
  so that they work when loading old files. I've reused the old field for the
  "before" range, so that at least something will be usable there.
* Some code cleanup is in order to make this code a bit nicer