Alembic support in the install_deps.sh script.
This is a rough-and-ready way of building and installing the Alembic
library and the necessary dependencies (excluding as much as possible
for now). The Alembic cmake files are very badly constructed and require
some annoying hacks atm. Eventually we might try to supply them with
patches or write our own build scripts for Alembic.