GPencil Editing: "Stroke Edit Mode"
Changed the way that the keybindings for the editing tools for strokes work,
to be less clunky to work with.

Now, instead of holding down DKEY, you instead enter "Stroke Edit Mode" (via
a toggle button in the "Edit Strokes" button in the toolbar). This enables a
keymap (Grease Pencil Stroke Edit Mode) which is only active when this setting
is enabled; the keybindings in this keymap are designed so that pressing the
normal hotkeys will affect the strokes as expected (instead of whatever else
may be in the viewport at the time instead).