GPencil Draw Operator: Implemented proper method to delay start of drawing
When running GPencil drawing operators from the toolbar or from a (pie) menu,
drawing operations shouldn't start immediately, but rather, only when users
actually click to start drawing. This is because the mouse is often not in a
suitable location for this to work when the operators get invoked from such
UI controls.

Previously, we used a hack which just detected certain types of events. This worked
ok for toolbars, but not for menus. Now, we've got a proper property to enable this
behaviour instead.