Proper implementation for "Volumetric Strokes" drawing
Due to the enthusiastic reaction from the community, I've decided to make this
a proper drawing technique. This commit separates the drawing code for this
out into its own method, and reduces the amount of costly GL read/write calls used
by performing more of this using Blender's internal math library. As a result,
it seems (at least in my limited tests) to be running much faster :)

TODO:
- Configuration options for controlling when this is used in combination
  with the other drawing techniques
- Use for stroke drawing too when enabled?
- Allow setting stroke sizes to be much thicker (and also set a better base-size)