GPencil Layer Lock - Does not prevent active frame from changing anymore
Loosened the restrictions on when the current frame can change,
by removing the restriction that the current frame on locked layers
cannot change (we already have the Frame Locking functionality for
that purpose instead). This makes it easier to animate (especially
with PropEdit enabled, when you need to lock layers to prevent
backgrounds and similar things from changing unexpectededly).