Depsgraph: Playback doesn't crash now
The time source is expected to be in the root depsgraph node for now,
so we should add it to there instead of the scene node.

This raises some questions:

- Who don't we have Scene ID node in the graph so far?
- Do we really need a time as an operation node, or it's something
  more up to the EvaluationContext.

Later one raises the bigger design question: what should we try to
do with the depsgraph and what is we're gonna to solve with the
evaluation context?

The answer here should be: everything which describes the (sub)graph
state should be in the evaluation context. This way this context will
finally become some key which totally corresponds to the (sub)scene
state.