Depsgraph: Add ubernode which evaluates the whole object
This way we mimic old depsgraph behavior where each object is effectively
a single node in the dependency graph which takes care of all the updates.

But actually this node only takes care of the modifiers, expecting that the
constraints and so are solved by the local object transform node.

Not sure how much valid the assumption at this state, but it's something
to aim to.