Remove 'topochange' from EDBM.
In future, there may be several tools relying on topo hash, each one making
updates to it when it needs. Hence, such thing should absolutely not be held
by mesh data, but (in this case) by Key data.

Also made some reorganization and renaming a bit.

Note: this code compiles, but for now cannot really be tested, since scratch
is not enabled...