Reshuffling of extra tool settings
Previously it was necessary to expose the editmode and stroke placement settings
at the bottom of the settings panel when it was displayed in the Sequencer, as
the sequencer doesn't have a toolbar for hosting those settings.

This commit moves those out into a separate panel, which can be included in other
editors too. This applies to editors which have a toolshelf for displaying these
settings too (e.g. nodes editor), where the toolshelf is often collapsed, making it
valuable to be able to set these settings from more than one place.

(While not great from a long-term standpoint, such duplication is a "good enough"
solution in the meantime, until we figure out a better solution)