WIP convenience feature: New frames are created when editing strokes on frames without existing keys
This feature makes it easier to do pose-to-pose animation using the strokes editing
stuff, as you don't have to remember to deliberately duplicate an existing frame
so that your edits end up on a new frame.

There is the possibility though that you didn't intend to create a new frame,
in which case this option will be more of a pain. Let's try this out for a bit
before deciding whether it should stay or go.