More mingw64 and c++11 fixes - hypot() is not defined
Fix is from the following thread(s):
* http://bugs.python.org/issue11566
* http://boost.2283326.n4.nabble.com/Boost-Python-Compile-Error-s-GCC-via-MinGW-w64-td3165793.html#a3166760

I haven't managed to fix this for all files here yet (this only fixes half of the files
for the freestyle module), so in the meantime, it'll be necessary to disable
the freestyle module. The same goes for Audaspace, which fails on the C APi currently.