Depsgraph: Add check for whether graph was properly tagged for rebuild
Currently it's more like a preliminary check to catch obvious errors based on
comparing number of operations in current depsgraph and correct one. Ideally
relations and nodes themselves should be checked deeper,but that's NP-complex
problem which probably doesn't worth solving for now.

This check is commented out by default, uncomment it in BKE_scene_update_tagged
if needed.