Depsgraph: Removing constraint should rebuild relations
TODO: We need to decide if we want to try supporting partial graph
updates or we'll just rebuild the whole graph when something changes.

Perhaps we'd better stick to the later for the simplicity and make
sure depsgraph build is not that slow.

Hunt for the missing relation updates continues.