Depsgraph: Use explicit guarded allocation macros
This way we always are keeping track on memory usage of the depsgraph
(as opposite to previous: only when GXX guarded allocator is enabled,
which isn't enabled i.e. for release builds).

Allows us to track memory leaks earlier, plus allows us to keep an eye
on the memory usage of the graph itself.