Depsgraph: Fallback to boost's function binding if C++11 is disabled
C++11 appears to be quite tricky to support in a clean fashion with all
the libraries we want blender to compile with. Making those libraries
friendly with C++11 could take a bit of time, and meanwhile instead of
having hacks all over the code we'll be able to just use boost.

This would help avoiding merging crappyness to the master branch, but
idea;;y we'll need to make all the libs being nice with C++11 and zap
boost from the sources.

Consider this as a safety backup plan commit.