Tweaks to where placeholder ops for geometry eval go
The original intention here was that operation nodes for geometry would go
on the geometry datablocks. Currently, these are added to the object datablock
instead so that it works with old files.

The main result of this change is a lot of the relation-building code now works
correctly again, since there's only a single node again in the geometry component there.