Get rid of verify_entry_exit_operations()
This was a holdover from the initial implementation, when we tried to do
everything in a single pass, and we couldn't be sure if a component's
start/end operations had been created yet (i.e. since some components
may have only been partially created).

Following discussion with Sergey, we've concluded that this is redundant