Depsgraph: Use layers visibility passed from scene_update_for_newframe
This way we can easily deal with the cases when render engine requests for
a different set of layers to be updated without need make bigger changes to
the outside of the dependency graph.

TODO: It is now possible to have some missing scene updates when mixing
update from the viewport and render engine. This isn't difficult to solve
but wouldn't be able to wrap up the patch today.