Code Cleanup: Disambiguate the "ComponentKey" name
ComponentKey was mostly used to refer to the lookup keys used when specifying
a relevant node for use when creating a relationship, while the secondary
usage of this was in IDDepsNodes as the keys for the component dict. While there
is technically no problem with having this, in practice it was too confusing.

So, now "ComponentKey" is only for looking up nodes, while "ComponentIDKey" is
used internally to identify keys.