Depsgraph: Use spin lock in DepsgraphDebug
Amount of oepration was really small so using mutex was a total overkill.

On the other hand, this code didn't really run on normal scene update
(which i understood after doing the changes..) so it's not like you can
expect any visible changes from this. More like "nice to have".