Depsgraph: Move pose evaluation callbacks from temporary stub to blenkernel
Now all the callback functions used by dependency graph are moved to dedicated
armature_update.c file in blenkernel.

There are also some spline IK functions moved from armature.c to the new file
which isn't strictly related on the original change but without any file in
blenkernel using functions from the new armature_update there'll be a linking
error. It also seems not so crazy idea to concentrate all the pose evaluation
in a single file, so overall it's not so much bad. Just annoying stupidnes
of compiler forced to do such a move now.