Depsgraph: Inserting keyframes for the first time now rebuilds the depsgraph
Since animation playback now works only if there's a link to the time source node,
it is necessary to rebuild the depsgraph when animation gets added to a datablock
for the first time (i.e. when keyframing on a previously unkeyed datablock) to
ensure that such a link gets added. It is not however needed to keep adding this
on subsequent keyframing attempts.