Depsgraph: Make freestyle compatible with C++11
This commit makes code aware of auto_ptr being deprecated in C++11 and
makes it so unique_ptr is used instead. There is now utility header file
for this which defines AutoPtr to either unique or auto pointer.

The only unclear thing is how to cast unique_ptr to pointer of another
type. getting some weird compiler issue with that. So if some of the
C++ guys around can have a look that'd help a lot. For now using really
dumb approach to solve that.

This doesn't mean freestyle works with new depsgraph -- there is some
missing scene update which makes lines being misaligned with the scene.
Will have a look later into this issue. But with --legacy-depsgraph
command line argument freestyle still seems to work.