Framebuffer effects:
This includes offscreen processing for the 3D viewport and OpenGL rendering.

Included effects are a depth of field shader and a Screen Space Ambient
Occlusion shader.

The system is simplistic and duplicates some code. I am in the process of making abstractions
so the effects can be used in a node system, but we can probably give users
access to the early system right now.

Reviewers: sergey, campbellbarton, jwilkins, merwin

Differential Revision: https://developer.blender.org/D1092