Depsgraph: Initial support of dupligroups
This is an attempt to bring the dupligroups support at least on the same
level as they used to work in the old dependency graph. This means we don't
bother with local storage, time offset etc for now and building the group
graph in the same scene graph (as opposed t ideal sub-graph),

The reason for that is that it'll allow to have working system sooner,
and even such a simplified approach seems to have issues with updates.