Depsgraph: Fix missing curve path in certain situations
Basically this is a quick re-implementation of evaluation flags we
had in the old dependency graph. I know it is not the real solution
but it's not so bad as well and would allow us continue nailing all
the other bugs with current granularity level before we go more
advanced.