Big revert of all things pointcache, old code is impossible to work with.
The Alembic caching must be integrated with Blender as a new, entirely
independent feature. The old pointcache code can not be surgically
removed without major damage to vital features. Instead of wasting more
time trying to deal with an outdated complicated system, a better choice
is implementing a new caching system next to it and if possible phase
out the old one at some point ...