Depsgraph: Temp solution for calculate motion paths
Re-enable armature handing from old object_handle_update.

For this to work fully we'll need to keep track of ob->recalc, which
is also seems to be quite needed for proper re-tagging after relations
are built anyway.

And in any case, for features like this seems having single function
to update given object is not totally terribly idea.